sorm: poprawki, obsługa definiowania kolumn

+++ wxgtd/model/sorm.py
@@ -1,33 +1,21 @@
 # -*- coding: utf-8 -*-
 
 """
-Prosty ORM
+very Simple ORM
 """
 
 __author__ = "Karol Będkowski"
-__copyright__ = "Copyright (c) Karol Będkowski, 2009-2013"
-__version__ = "2011-05-15"
+__copyright__ = "Copyright (c) Karol Będkowski, 2013"
+__version__ = "2013-03-05"
 
 
 import logging
 import sqlite3
 import locale
 import itertools
+#import types
 
 _LOG = logging.getLogger(__name__)
-
-
-class _Singleton(object):
-	def __new__(cls, *args, **kwarg):
-		instance = cls.__dict__.get('__instance__')
-		if instance is None:
-			instance = object.__new__(cls)
-			instance._init(*args, **kwarg)
-			cls.__instance__ = instance
-		return instance
-
-	def _init(self, *args, **kwarg):
-		pass
 
 
 class _CursorWrapper(object):
@@ -44,8 +32,17 @@ class _CursorWrapper(object):
 		self._cursor.close()
 
 
-class DbConnection(_Singleton):
+class DbConnection(object):
 	""" Singleton for all db connections """
+
+	_instance = None
+
+	def __new__(cls, *args, **kwarg):
+		if cls._instance is None:
+			cls._instance = object.__new__(cls)
+			cls._instance._init(*args, **kwarg)
+		return cls._instance
+
 	def _init(self):
 		self._connection = None
 
@@ -85,22 +82,91 @@ class DbConnection(_Singleton):
 		_LOG.error("DbConnection.execute ERROR: not connected")
 
 
+class Column(object):
+	"""Column definition"""
+	def __init__(self, name=None, value_type=None, default=None,
+			primary_key=False):
+		self.name = name
+		self.value_type = value_type  # convert type, None=don't change
+		self.default = default
+		self.primary_key = primary_key
+
+	def __repr__(self):
+		return ' '.join(map(str, ('<Column', self.name, self.value_type,
+			self.default, self.primary_key, '>')))
+
+	def from_database(self, value):
+		if self.value_type is None:
+			return value
+		return self.value_type(value)
+
+	def to_database(self, value):
+		return value
+
+
+class _MetaModel(type):
+	"""MetaModel for Model class"""
+	def __new__(mcs, name, bases, dict_):
+		if '_primary_keys' not in dict_:
+			dict_['_primary_keys'] = []
+		fields = dict_['_fields']
+		if isinstance(fields, (list, tuple, set)):
+			rfields = {}
+			for field in fields:
+				if isinstance(field, Column):
+					if not field.name:
+						raise TypeError('No field name')
+					rfields[field.name] = field
+				elif isinstance(field, (str, unicode)):
+					rfields[field] = Column(name=field)
+				else:
+					raise TypeError('Invalid column definition %r'
+							% field)
+			dict_['_fields'] = rfields
+		dict_['_primary_keys'] = set(dict_['_primary_keys'])
+		return type.__new__(mcs, name, bases, dict_)
+
+	def __init__(mcs, name, bases, dict_):
+		type.__init__(mcs, name, bases, dict_)
+		fields = dict_['_fields']
+		primary_keys = dict_['_primary_keys']
+		if isinstance(fields, dict):
+			for key, value in fields.iteritems():
+				if isinstance(value, Column):
+					if not value.name:
+						value.name = key
+					if value.primary_key:
+						if value.name not in primary_keys:
+							primary_keys.add(value.name)
+				elif isinstance(value, (str, unicode)):
+					fields[key] = Column(name=value,
+							primary_key=(value in primary_keys))
+				else:
+					fields[key] = Column(name=key)
+
+
 class Model(object):
-	"""docstring for Model"""
+	"""Base class for all objects"""
+	__metaclass__ = _MetaModel
 
 	_table_name = None
-	_fields = {}  # map class property -> db field (optional)
-	_primary_keys = []  # list of primary keys properties
+	_fields = {}  # map class property -> None | db field name (str) | Column()
+	_primary_keys = set()  # list of primary keys properties
+
+	def __new__(cls, *args, **kwarg):
+		instance = object.__new__(cls, *args, **kwarg)
+		# prepare instance
+		for key, column_def in cls._fields.iteritems():
+			instance.__dict__[key] = column_def.default
+		return instance
 
 	def __init__(self, **kwargs):
 		super(Model, self).__init__()
-		for key in self._fields.keys():
-			setattr(self, key, None)
 		for key, val in kwargs.iteritems():
 			if key in self._fields:
 				setattr(self, key, val)
 
-	def __str__(self):
+	def __repr__(self):
 		res = ['<', self.__class__.__name__,
 				'TABLE=%r' % self._table_name]
 		res.extend('%r=%r' % (key, val) for key, val
@@ -118,7 +184,9 @@ class Model(object):
 		with DbConnection().get_cursor() as cursor:
 			cursor.execute(sql, query_params)
 			for row in cursor:
-				yield cls(**row)
+				values = dict((key, cls._fields[key].from_database(val))
+						for key, val in dict(row).iteritems())
+				yield cls(**values)
 
 	@classmethod
 	def get(cls, **where):
@@ -141,6 +209,15 @@ class Model(object):
 				query_params)
 		with DbConnection().get_cursor() as cursor:
 			cursor.execute(sql, query_params)
+			rowid = cursor.lastrowid
+			# update from db
+			sql, query_params = self.__class__._create_select_query(limit=1,
+					rowid=rowid)
+			cursor.execute(sql, query_params)
+			row = cursor.fetchone()
+			for key, val in dict(row).iteritems():
+				if key in self._fields:
+					setattr(self, key, self._fields[key].from_database(val))
 
 	def update(self):
 		"""Update current object.
@@ -167,9 +244,9 @@ class Model(object):
 				distinct=None, where_stmt=None, group_by=None, **where):
 		"""Prepare select query for given parameters """
 		sqls = ['SELECT', ("DISTINCT" if distinct else "")]
-		# use "AS" if object fiels is diferrent from table column
-		sqls.append(', '.join(((field + " AS " + key)
-				if (field and field != key) else key)
+		# use "AS" if object fields is different from table column
+		sqls.append(', '.join(((field.name + " AS " + key)
+				if (field.name and field.name != key) else key)
 				for key, field in cls._fields.iteritems()))
 		sqls.append('FROM "%s"' % cls._table_name)
 		query_params = []
@@ -180,7 +257,9 @@ class Model(object):
 				where_params.append(where_stmt)
 			if where:
 				for column, value in where.iteritems():
-					where_params.append("%s=?" % (cls._fields[column] or column))
+					column_name = (cls._fields[column].name or column
+							if column in cls._fields else column)
+					where_params.append("%s=?" % column_name)
 					query_params.append(value)
 			sqls.append(' AND '.join(where_params))
 		if group_by:
@@ -197,12 +276,10 @@ class Model(object):
 
 	def _create_update_stmt(self):
 		"""Prepare sql stmt and parameters for update current object."""
-		assert bool(self._primary_keys)
-		values = [((field or key), getattr(self, key))
-				for key, field in self._fields.iteritems()
-				if key not in self._primary_keys]
-		pkeys = [(self._fields[key] or key, getattr(self, key))
-				for key in self._primary_keys]
+		pkeys = self._get_pkey_values()
+		if not pkeys or not all(pkey[1] for pkey in pkeys):
+			raise RuntimeError('Missing primary keys')
+		values = self._get_attr_values(False)
 		sql = ['UPDATE', self._table_name, 'SET']
 		sql.append(', '.join(val[0] + "=?" for val in values))
 		sql.append('WHERE')
@@ -213,8 +290,7 @@ class Model(object):
 
 	def _create_save_stmt(self):
 		"""Prepare sql for save (insert) new object"""
-		values = [(field or key, getattr(self, key)) for key, field
-				in self._fields.iteritems()]
+		values = self._get_attr_values(True)
 		sql = ['INSERT INTO', self._table_name, '(']
 		sql.append(', '.join(val[0] for val in values))
 		sql.append(') VALUES (')
@@ -226,11 +302,28 @@ class Model(object):
 
 	def _create_delete_stmt(self):
 		"""Prepare sql for delete object"""
-		assert bool(self._primary_keys)
-		pkeys = [(self._fields[key] or key, getattr(self, key))
-				for key in self._primary_keys]
+		pkeys = self._get_pkey_values()
+		if not pkeys or not all(pkey[1] for pkey in pkeys):
+			raise RuntimeError('Missing primary keys')
 		sql = ['DELETE FROM', self._table_name, 'WHERE']
 		sql.append(' AND '.join(val[0] + "=?" for val in pkeys))
 		sql = ' '.join(sql)
 		query_params = [val[1] for val in pkeys]
 		return sql, query_params
+
+	def _get_pkey_values(self):
+		'''Get [(primary key, value)] '''
+		fields = self._fields
+		return [(fields[key].name or key,
+				fields[key].to_database(getattr(self, key)))
+				for key in self._primary_keys]
+
+	def _get_attr_values(self, with_pkeys=False):
+		'''Get [(column, value)] for all columns with primary key or without'''
+		if with_pkeys:
+			return [(field.name or key, field.to_database(getattr(self, key)))
+					for key, field in self._fields.iteritems()]
+		else:
+			return [((field.name or key), field.to_database(getattr(self, key)))
+					for key, field in self._fields.iteritems()
+					if key not in self._primary_keys]
